OGM-1002 Move Redis
