+++ integrationtest/pom.xml
@@ -173,6 +173,11 @@
             <artifactId>hibernate-ogm-mongodb</artifactId>
             <scope>provided</scope>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>hibernate-ogm-redis</artifactId>
+            <scope>provided</scope>
+        </dependency>
 
         <dependency>
             <groupId>org.wildfly.arquillian</groupId>
@@ -267,6 +272,38 @@
             </build>
         </profile>
 
+        <profile>
+            <id>run-redis-wildfly-tests</id>
+            <!-- To run the tests you need to define the
+                environment property: REDIS_HOSTNAME -->
+            <activation>
+                <property>
+                    <name>env.REDIS_HOSTNAME</name>
+                </property>
+            </activation>
+            <build>
+                <plugins>
+                    <plugin>
+                        <artifactId>maven-failsafe-plugin</artifactId>
+                        <executions>
+                            <execution>
+                                <id>redis-integration-tests</id>
+                                <goals>
+                                    <goal>integration-test</goal>
+                                    <goal>verify</goal>
+                                </goals>
+                                <configuration>
+                                    <redirectTestOutputToFile>true</redirectTestOutputToFile>
+                                    <includes>
+                                        <include>**/redis/*IT.java</include>
+                                    </includes>
+                                </configuration>
+                            </execution>
+                        </executions>
+                    </plugin>
+                </plugins>
+            </build>
+        </profile>
 
         <profile>
             <id>run-cassandra-wildfly-tests</id>
